build: Add "make dist"
For creating a tarball without .git directory easier.  The tarball will
include version.ent and LFS-RELEASE files so the book rendered from it
won't show "unknown" for all dates.

+++ git-version.sh
@@ -17,6 +17,11 @@ echo "<!ENTITY % systemd \"$SYSTEMD\">"  >> conditional.ent
 if ! git status > /dev/null; then
     # Either it's not a git repository, or git is unavaliable.
     # Just workaround.
+
+	if [ -e LFS-RELEASE ]; then
+		exit 0
+	fi
+
     echo "<![ %sysv; ["                                    >  version.ent
     echo "<!ENTITY version           \"unknown\">"         >> version.ent
     echo "]]>"                                             >> version.ent
@@ -65,3 +70,5 @@ echo "<!ENTITY version          \"$versiond\">"            >> version.ent
 echo "]]>"                                                 >> version.ent
 echo "<!ENTITY releasedate       \"$full_date\">"          >> version.ent
 echo "<!ENTITY copyrightdate     \"1999-$year\">"          >> version.ent
+
+[ -z "$DIST" ] || echo $version > "$DIST"
